Move HH:MM:SS duration formatting to a reusable utility
